+++ tests/parser/scripts/comments.py
@@ -5,24 +5,24 @@ def f(a: int):
     # 1
     if a:
         # 2
-        b = (a + 2)
+        b = a + 2
         # 3
-    elif (a + 1):
+    elif a + 1:
         # 4
-        b = (a + 1)
+        b = a + 1
         # 5
     else:
         # 6
         b = a
         # 7
     # 8
-    c = (b * 3)
+    c = b * 3
     # 9
     return c
 
 # example header1
 # example header2
-def g(b : int):
+def g(b: int):
     # 10
     while b:
         # 11
@@ -35,7 +35,7 @@ def g(b : int):
             # 15
             print(i)
             # 16
-            if (i == 4):
+            if i == 4:
                 # 17
                 k = 3
                 # 18
@@ -55,16 +55,15 @@ def h(d: int):
     # 26
     if d:
         # 27
-        e = (d + 2)
+        e = d + 2
         # 28
+    elif d + 1:
+        # 29
+        e = d + 1
+        # 30
     else:
-        if (d + 1):
-            # 29
-            e = (d + 1)
-            # 30
-        else:
-            # 31
-            e = d
-            # 32
+        # 31
+        e = d
+        # 32
     # 33
     return e