+++ CHANGELOG.md
@@ -129,6 +129,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+-   #2008 : Remove dependency on `astunparse` package.
 -   #1920 : Add a maximum version for NumPy.
 -   #1836 : Move `epyccel` module to `pyccel.commands.epyccel` and add support for shortcut import `from pyccel import epyccel`.
 -   #1720 : functions with the `@inline` decorator are no longer exposed to Python in the shared library.
@@ -172,7 +173,7 @@ All notable changes to this project will be documented in this file.
 
 ### Deprecated
 
--   #1820 : Deprecated unused decorator `@lambdify`
+-   #2008 : Remove support for Python 3.8.
 -   #1786 : Remove support for `real` and `integer` as type annotations.
 -   #1812 : Stop allowing multiple main blocks inside a module.
 -   \[INTERNALS\] Remove property `ast.basic.TypedAstNode.precision`.